fix: moved userNameInitials to own helper file to use it elsewhere, couple refactorings
